merlin-scraper_backend/requirements.txt

autopep8==2.3.2
blobfile==3.0.0
Flask==3.1.0
flask-cors==5.0.1
gensim==4.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.14.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
huggingface-hub==0.31.1
numpy==1.26.4
openai==1.78.0
pillow==11.1.0
prompt_toolkit==3.0.50
scikit-learn==1.6.1
selenium==4.32.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
pymongo
dotenv
pymupdf
bs4
pandas
ics